REUP Model, DAO, Service and CREATE UI

+++ src/main/java/com/mdk/dao/IDetailOrdersItemDAO.java
@@ -3,5 +3,5 @@ package com.mdk.dao;
 import com.mdk.models.DetailOrdersItem;
 
 public interface IDetailOrdersItemDAO {
-    DetailOrdersItem findOneByOrderId(Long id);
+    DetailOrdersItem findOneByOrderId(int id);
 }
