+++ src/main/java/com/mdk/models/Cart.java
@@ -1,22 +1,22 @@
 package com.mdk.models;
 
 public class Cart extends AbstractModel<Cart> {
-    private Long userId;
-    private Long storeId;
+    private int userId;
+    private int storeId;
 
-    public Long getUserId() {
+    public int getUserId() {
         return userId;
     }
 
-    public void setUserId(Long userId) {
+    public void setUserId(int userId) {
         this.userId = userId;
     }
 
-    public Long getStoreId() {
+    public int getStoreId() {
         return storeId;
     }
 
-    public void setStoreId(Long storeId) {
+    public void setStoreId(int storeId) {
         this.storeId = storeId;
     }
 }
