+++ src/main/java/com/mdk/services/IDetailOrdersItemService.java
@@ -3,6 +3,7 @@ package com.mdk.services;
 import com.mdk.models.DetailOrdersItem;
 
 public interface IDetailOrdersItemService {
-    DetailOrdersItem findOneByOrderId(Long id);
+    DetailOrdersItem findOneByOrderId(int id);
+
 
 }
